--- start.dat
@@ -1,32 +1,0 @@
-  0.40000000000000002        478.39118947478329        842.48164824095784        1.0000000000000000      -0.98113727569580078     
-  0.40000000000000002        260.27208861074138        458.45047406776484        1.0000000000000000      -0.90377902984619141     
-  0.40000000000000002        168.00838435551762        295.95957248363834        1.0000000000000000       0.51726102828979492     
-  0.40000000000000002        263.32210994828966        464.10707435458465        1.0000000000000000      -0.52854311466217041     
-  0.40000000000000002        341.99723115339685        602.22919633654328        1.0000000000000000      -0.23999321460723877     
-  0.40000000000000002        516.23705539358457        909.29803222506177        1.0000000000000000      -0.32966625690460205     
-  0.40000000000000002        278.92146431513237        491.07672729433546        1.0000000000000000        9.5226764678955078E-003
-  0.40000000000000002        571.23490342596085        1005.9273884297247        1.0000000000000000       -7.6891183853149414E-003
-  0.40000000000000002        579.91428593278704        1021.4986178170474        1.0000000000000000      -0.58524906635284424     
-  0.40000000000000002        1007.7287826313469        1774.8741668047990        1.0000000000000000      -0.92209053039550781     
-  0.40000000000000002        1185.2454066097598        2087.8384791492904        1.0000000000000000      -0.63333547115325928     
-  0.40000000000000002        1090.7321519353020        1921.5573544328283        1.0000000000000000      -0.41920936107635498     
-  0.40000000000000002        88.824648465386460        156.72170975692634        1.0000000000000000       0.24921035766601562     
-  0.40000000000000002        1082.8758204056699        1907.3105247037781        1.0000000000000000      -0.44980108737945557     
-  0.40000000000000002        945.08045803945197        1664.6851264107922        1.0000000000000000       -3.6953330039978027E-002
-  0.40000000000000002        150.25773787991611        264.34827433262802        1.0000000000000000       0.33278489112854004     
-  0.40000000000000002        983.78139549742650        1733.1094110965807        1.0000000000000000       0.47448670864105225     
-  0.40000000000000002        412.45054580228873        726.54863827797260        1.0000000000000000       0.35641741752624512     
-  0.40000000000000002        1005.4026953906404        1771.3301139299897        1.0000000000000000       0.76085937023162842     
-  0.40000000000000002        476.24440451234312        839.14907849726694        1.0000000000000000      -0.11733186244964600     
-  0.40000000000000002        606.83145746874573        1069.1403381277721        1.0000000000000000       0.17465484142303467     
-  0.40000000000000002        190.07538584918811        334.87898624772117        1.0000000000000000      -0.23849177360534668     
-  0.40000000000000002        417.38011812003185        735.24561149891861        1.0000000000000000      -0.56056022644042969     
-  0.40000000000000002        470.09673307211972        828.29258821134670        1.0000000000000000      -0.27763199806213379     
-  0.40000000000000002        455.95032004146873        803.29104208999627        1.0000000000000000       0.84378325939178467     
-  0.40000000000000002        1107.8380884566352        1951.3790146449892        1.0000000000000000      -0.87953472137451172     
-  0.40000000000000002        76.737613560752351        134.87132539536501        1.0000000000000000       0.67638504505157471     
-  0.40000000000000002        871.44559461209246        1535.0515476193523        1.0000000000000000       0.87100565433502197     
-  0.40000000000000002        471.64329368523880        831.10871661938199        1.0000000000000000       0.14747166633605957     
-  0.40000000000000002        753.70798086386276        1327.6976205635067        1.0000000000000000       0.62948060035705566     
-  0.40000000000000002        1105.0654996809524        1946.6252194457400        1.0000000000000000       0.82120621204376221     
-  0.40000000000000002        852.95589851349871        1502.7998472776528        1.0000000000000000      -0.79872477054595947
