+++ tests/test_fields.py
@@ -1,9 +1,10 @@
 import logging
 import unittest
 
+import numpy as np
 from numpy.testing import assert_almost_equal
 
-from neat.fields import Stellna, StellnaQS
+from neat.fields import Simple, Stellna, StellnaQS
 
 logging.basicConfig(level=logging.INFO)
 logger = logging.getLogger(__name__)
@@ -68,3 +69,105 @@ class NEATtests(unittest.TestCase):
             -0.018692578813516082,
             decimal=10,
         )
+
+    def setUp(self):
+        self.simple_object = Simple(
+            wout_filename="/home/rodrigo/NEAT/examples/inputs/wout_ARIESCS.nc",
+            B_scale=1.0,
+            Aminor_scale=1.0,
+            multharm=3,
+        )
+
+    def test_simple_single_particle_tracer(self):
+        # Set up input parameters for the single particle tracer
+        Tracy = self.simple_object.tracy
+        Rmajor = self.simple_object.Rmajor
+        Simple = self.simple_object.simple
+        charge = 1.0
+        mass = 1.0
+        Lambda = 0.5
+        vpp_sign = 1.0
+        energy = 1.0
+        r_initial = 1.0
+        theta_initial = 0.0
+        phi_initial = 0.0
+        nsamples = 100
+        tfinal = 1.0
+
+        result = self.simple_object.simple_single_particle_tracer(
+            Tracy,
+            Rmajor,
+            Simple,
+            charge,
+            mass,
+            Lambda,
+            vpp_sign,
+            energy,
+            r_initial,
+            theta_initial,
+            phi_initial,
+            nsamples,
+            tfinal,
+        )
+
+        expected_shape = (15,)
+        self.assertEqual(result[0].shape, expected_shape)
+
+    def test_simple_ensemble_particle_tracer(self):
+        # Set up input parameters for the ensemble particle tracer
+
+        Tracy = self.simple_object.tracy
+        Rmajor = self.simple_object.Rmajor
+        Simple = self.simple_object.simple
+
+        nlambda_trapped = 10
+        nlambda_passing = 10
+        r_initial = 0.1
+        r_max = 0.9
+        ntheta = 20
+        nphi = 20
+        nthreads = 4
+        vparallel_over_v_min = -0.3
+        vparallel_over_v_max = 0.3
+        npoiper = 100
+        npoiper2 = 100
+        nper = 1000
+        nsamples = 2000
+        tfinal = 0.001
+        r_initial = 0.12
+        energy = 3.52e6
+        charge = 2
+        mass = 4
+
+        result = self.simple_object.simple_ensemble_particle_tracer(
+            Tracy,
+            Rmajor,
+            Simple,
+            charge,
+            mass,
+            energy,
+            nlambda_trapped,
+            nlambda_passing,
+            r_initial,
+            r_max,
+            ntheta,
+            nphi,
+            nsamples,
+            tfinal,
+            nthreads,
+            vparallel_over_v_min,
+            vparallel_over_v_max,
+            npoiper,
+            npoiper2,
+            nper,
+        )
+
+        # Perform assertions on the result
+
+        # Example assertion: check if time array has the correct shape
+        expected_shape = (nsamples,)
+        self.assertEqual(result[0].shape, expected_shape)
+
+
+if __name__ == "__main__":
+    unittest.main()
